+++ sample/src/jvmMain/kotlin/Main.kt
@@ -20,12 +20,15 @@ import io.gladed.watchable.store.cannot
 import io.ktor.application.Application
 import io.ktor.application.call
 import io.ktor.application.install
+import io.ktor.features.CachingHeaders
 import io.ktor.features.CallLogging
 import io.ktor.features.ContentNegotiation
 import io.ktor.features.StatusPages
 import io.ktor.html.respondHtml
+import io.ktor.http.CacheControl
 import io.ktor.http.ContentType
 import io.ktor.http.HttpStatusCode
+import io.ktor.http.content.CachingOptions
 import io.ktor.http.content.files
 import io.ktor.http.content.static
 import io.ktor.request.receiveParameters
@@ -37,11 +40,13 @@ import io.ktor.routing.route
 import io.ktor.routing.routing
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.flow.toList
+import kotlinx.coroutines.withContext
 import kotlinx.html.FormMethod
 import kotlinx.html.InputType
 import kotlinx.html.a
@@ -51,8 +56,10 @@ import kotlinx.html.div
 import kotlinx.html.form
 import kotlinx.html.h1
 import kotlinx.html.h2
+import kotlinx.html.h3
 import kotlinx.html.head
 import kotlinx.html.input
+import kotlinx.html.p
 import kotlinx.html.title
 import logic.Logic
 import model.Bird
@@ -67,24 +74,32 @@ const val LOCAL_HOST = "127.0.0.1"
 
 fun main() { Main.setup() }
 
+actual class Sample {
+    actual fun checkMe() = 42
+}
+
+actual object Platform {
+    actual val name: String = "JVM"
+}
+
 object Main {
     fun setup() {
+        Sample().checkMe()
+        Platform.name
         embeddedServer(Netty, port = SAMPLE_PORT, host = LOCAL_HOST, watchPaths = listOf("build/classes"),
             module = Application::setup)
             .start(wait = true)
     }
-
-    fun setup(app: Application, logic: Logic) = app.setup(logic)
 }
 
 fun Application.setup() {
     val dataDir = File(".data")
     val logic = Adapter.createLogic(coroutineContext, dataDir)
-    setup(logic)
+    bind(logic)
 }
 
 @UseExperimental(FlowPreview::class)
-fun Application.setup(logic: Logic) {
+fun Application.bind(logic: Logic) {
     val routes = Routes(logic)
     val currentDir = File(".").absoluteFile
     environment.log.info("Current directory: $currentDir")
@@ -98,7 +113,9 @@ fun Application.setup(logic: Logic) {
         }.firstOrNull { it.isDirectory }?.absoluteFile ?: error("Can't find 'web' folder for this sample")
 
     suspend fun <T> withLogic(func: suspend Logic.Scoped.() -> T): T =
-        coroutineScope { with(logic.scoped(this)) { func() } }
+        withContext(Dispatchers.Default + Job()) {
+            logic.scoped(this).func()
+        }
 
     install(ContentNegotiation) {
         register(ContentType.Application.Json, KotlinSerializationConverter()) {
@@ -113,10 +130,16 @@ fun Application.setup(logic: Logic) {
         }
     }
 
+    install(CachingHeaders) {
+        options {
+            //                ContentType.Text.CSS -> CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 24 * 60 * 60))
+            CachingOptions(CacheControl.NoCache(CacheControl.Visibility.Public))
+        }
+    }
+
     install(CallLogging) {
         level = Level.INFO
     }
-
 
     routing {
         with(routes) {
@@ -126,7 +149,8 @@ fun Application.setup(logic: Logic) {
         }
         get("/") {
             val topBirds = withLogic {
-                birds.keys().take(10).map { birds.get(it) }.toList()
+                birds.keys().take(10).map {
+                    birds.get(it) }.toList()
             }
             call.respondHtml {
                 head {
@@ -146,7 +170,8 @@ fun Application.setup(logic: Logic) {
 //                            +"require(['/static/sample.js'], function(js) { js.helloWorld('Hi'); });\n"
 //                        }
 //                    }
-                    h2 { +"Birds" }
+                    h1 { +"Birds" }
+                    p { +"Some birds:" }
                     topBirds.forEach { bird ->
                         div {
                             a(href="/bird/${bird.id}") { +bird.name.value }
@@ -154,11 +179,13 @@ fun Application.setup(logic: Logic) {
                     }
 
                     form(action = "bird", method = FormMethod.post) {
-                        +"Name: "
-                        input(InputType.text, name = "name")
+                        h3 { +"Create a bird" }
+                        input(InputType.text, name = "name") {
+                            placeholder = "name of new bird"
+                        }
                         +" "
                         input(InputType.submit) {
-                            value = "Create Bird"
+                            value = "Create"
                         }
                     }
                 }
@@ -177,11 +204,13 @@ fun Application.setup(logic: Logic) {
             call.respondRedirect("/bird/${bird.id}")
         }
 
-        // Delete a bird by ID
+        // $a bird by ID
         post("/bird/{birdId}/delete") {
-            withLogic {
+            val done = withLogic {
                 birds.remove(call.parameters["birdId"]!!)
+                coroutineContext[Job]
             }
+            environment.log.info("Done $done")
             call.respondRedirect("/")
         }
 
@@ -189,8 +218,8 @@ fun Application.setup(logic: Logic) {
         get("/bird/{birdId}") {
             withLogic {
                 birds.get(call.parameters["birdId"]!!).let { bird ->
-                    val myChirps = ops.chirpsForBird(bird.id).take(10).map {
-                        chirps.get(it)
+                    val related = ops.relatedChirps(bird.id).take(10).map { chirpKey ->
+                        chirps.get(chirpKey).let { chirp -> chirp to birds.get(chirp.from) }
                     }.toList()
                     val follows = bird.following.map { birds.get(it) }
 
@@ -202,30 +231,34 @@ fun Application.setup(logic: Logic) {
                             h1 { +bird.name.value }
                             form(action = "/bird/${bird.id}/delete", method = FormMethod.post) {
                                 input(InputType.submit) {
-                                    value = "Delete"
+                                    value = "Delete ${bird.name}"
                                 }
                             }
 
                             h2 { +"Chirps" }
-                            myChirps.forEach {
+                            related.forEach { (chirp, from) ->
                                 div {
-                                    +"At ${it.sentAt}, ${bird.name} chirped: ${it.text}"
+                                    +"${chirp.sentAt}, ${from.name}: ${chirp.text}"
                                     br
-                                    +"Reactions: ${it.reactions.values.sorted()}"
-                                    form(action = "/chirp/${it.id}/delete", method = FormMethod.post) {
-                                        input(InputType.submit) {
-                                            value = "Delete"
+                                    +"Reactions: ${chirp.reactions.values.sorted()}"
+                                    if (chirp.from == bird.id) {
+                                        form(action = "/chirp/${chirp.id}/delete", method = FormMethod.post) {
+                                            input(InputType.submit) {
+                                                value = "Delete chirp"
+                                            }
                                         }
                                     }
                                 }
                             }
 
                             form(action = "/chirp", method = FormMethod.post) {
+                                h3 { +"Chirp something from ${bird.name.value}:" }
                                 input(InputType.hidden, name = "bird") {
                                     value = bird.id
                                 }
-                                +"Text: "
-                                input(InputType.text, name = "text")
+                                input(InputType.text, name = "text") {
+                                    placeholder = "text to chirp"
+                                }
                                 +" "
                                 input(InputType.submit) {
                                     value = "Chirp"
@@ -233,22 +266,26 @@ fun Application.setup(logic: Logic) {
                             }
 
                             h2 { +"Follows" }
-                            follows.forEach {
+                            follows.forEach { following ->
                                 div {
-                                    a(href = "/bird/${it.id}") { +it.name.value }
+                                    a(href = "/bird/${following.id}") { +following.name.value }
                                     form(action = "/bird/${bird.id}/unfollow", method = FormMethod.post) {
                                         input(InputType.hidden, name = "bird") {
-                                            value = it.id
+                                            value = following.id
                                         }
                                         input(InputType.submit) {
-                                            value = "Unfollow"
+                                            value = "Unfollow ${following.name}"
                                         }
                                     }
                                 }
                             }
+
+
                             form(action = "/bird/${bird.id}/follow", method = FormMethod.post) {
-                                +"Name: "
-                                input(InputType.text, name = "name")
+                                h3 { +"Follow a new bird:" }
+                                input(InputType.text, name = "name") {
+                                    placeholder = " name of bird to follow"
+                                }
                                 +" "
                                 input(InputType.submit) {
                                     value = "Follow"