+++ sample/src/jvmMain/kotlin/Sample.kt
@@ -1,9 +1,0 @@
-
-actual class Sample {
-    actual fun checkMe() = 42
-}
-
-actual object Platform {
-    actual val name: String = "JVM"
-}
-