+++ sample/src/jvmMain/kotlin/logic/Operations.kt
@@ -5,9 +5,14 @@ import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.map
 import model.Chirp
 import io.gladed.watchable.store.Store
+import kotlinx.coroutines.flow.Flow
 import model.Bird
 
-/** Perform search operations based on available stores. */
+/**
+ * Perform search operations based on available stores.
+ *
+ * The default operations are probably not very scalable and should be overridden.
+ */
 @UseExperimental(FlowPreview::class)
 open class Operations(private val chirps: Store<Chirp>, private val birds: Store<Bird>) {
     /**
@@ -20,6 +25,18 @@ open class Operations(private val chirps: Store<Chirp>, private val birds: Store
             .map { chirps.get(it) }
             .filter { it.from == birdId }
             .map { it.id }
+
+
+    /**
+     * Chirps that related to birdId
+     */
+    open suspend fun relatedChirps(birdId: String): Flow<String> {
+        val bird = birds.get(birdId)
+        return chirps.keys()
+            .map { chirps.get(it) }
+            .filter { it.from == birdId || bird.following.contains(it.from) }
+            .map { it.id }
+    }
 
     /** Return all birds matching name. */
     open fun birdsWithName(name: String) =