+++ src/main/kotlin/io/gladed/watchable/store/Cannot.kt
@@ -18,3 +18,7 @@ package io.gladed.watchable.store
 
 /** Something cannot be done. */
 open class Cannot(doThis: String) : Exception(doThis)
+
+/** Attempt something, returning null if it throws [Cannot]. */
+inline fun <T : Any> maybe(attempt: () -> T): T? =
+    try { attempt() } catch (c: Cannot) { null }